refactor: 타입 중복 선언 제거
