--- apps/maeilwiki/src/common/types/NullableFormState.ts
@@ -1,3 +1,0 @@
-export type NullableFormState<FormState> = {
-  [K in keyof FormState]: FormState[K] | null;
-};
